# services/interbank-service/docs/swagger.yaml
definitions:
...
  dto.CurrencyCode:
    enum:
    - RSD
    - EUR
    - USD
    - CHF
    - JPY
    - AUD
    - CAD
    - GBP
    type: string
    x-enum-varnames:
    - CurrencyRSD
    - CurrencyEUR
    - CurrencyUSD
    - CurrencyCHF
    - CurrencyJPY
    - CurrencyAUD
    - CurrencyCAD
    - CurrencyGBP
  dto.ForeignBankId:
    properties:
      id:
        maxLength: 64
        type: string
      routingNumber:
        type: integer
    required:
    - id
    - routingNumber
    type: object
...
  dto.MonetaryValue:
    properties:
      amount:
        type: number
      currency:
        $ref: '#/definitions/dto.CurrencyCode'
    required:
    - amount
    - currency
    type: object
...
  dto.OtcNegotiationView:
    properties:
      id:
        $ref: '#/definitions/dto.ForeignBankId'
      offer:
        $ref: '#/definitions/dto.OtcOffer'
      status:
        description: ongoing | accepted | cancelled | expired
        type: string
      updatedAt:
        type: string
    type: object
  dto.OtcOffer:
    properties:
      amount:
        minimum: 1
        type: integer
      buyerAccountNumber:
        description: BuyerAccountNumber is a Banka-4 extension to the protocol's OtcOffer.
        type: string
      buyerId:
        $ref: '#/definitions/dto.ForeignBankId'
      lastModifiedBy:
        $ref: '#/definitions/dto.ForeignBankId'
      premium:
        $ref: '#/definitions/dto.MonetaryValue'
      pricePerUnit:
        $ref: '#/definitions/dto.MonetaryValue'
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      stock:
        $ref: '#/definitions/dto.StockDescription'
    required:
    - amount
    - buyerId
    - lastModifiedBy
    - pricePerUnit
    - sellerId
    - settlementDate
    - stock
    type: object
  dto.PeerContract:
    properties:
      amount:
        type: integer
      buyerId:
        $ref: '#/definitions/dto.ForeignBankId'
      createdAt:
        type: string
      exercisedAt:
        type: string
      id:
        $ref: '#/definitions/dto.ForeignBankId'
      negotiationId:
        $ref: '#/definitions/dto.ForeignBankId'
      premium:
        $ref: '#/definitions/dto.MonetaryValue'
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      status:
        type: string
      strikePrice:
        $ref: '#/definitions/dto.MonetaryValue'
      ticker:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.PublicStock:
    properties:
      sellers:
        items:
          $ref: '#/definitions/dto.PublicStockSeller'
        type: array
      stock:
        $ref: '#/definitions/dto.StockDescription'
    type: object
  dto.PublicStockSeller:
    properties:
      amount:
        type: integer
      seller:
        $ref: '#/definitions/dto.ForeignBankId'
    type: object
  dto.StockDescription:
    properties:
      ticker:
        type: string
    required:
    - ticker
    type: object
...
  errors.AppError:
    properties:
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  handler.AcceptPeerNegotiationRequest:
    type: object
  handler.CounterPeerNegotiationRequest:
    properties:
      amount:
        minimum: 1
        type: integer
      premium:
        type: number
      premiumCurrency:
        maxLength: 8
        type: string
      priceCurrency:
        maxLength: 8
        type: string
      pricePerStock:
        type: number
      settlementDate:
        type: string
    required:
    - amount
    - premium
    - premiumCurrency
    - priceCurrency
    - pricePerStock
    - settlementDate
    type: object
  handler.CreatePeerNegotiationRequest:
    properties:
      accountNumber:
        type: string
      amount:
        minimum: 1
        type: integer
      premium:
        type: number
      premiumCurrency:
        maxLength: 8
        type: string
      priceCurrency:
        maxLength: 8
        type: string
      pricePerStock:
        type: number
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      ticker:
        maxLength: 16
        type: string
    required:
    - accountNumber
    - amount
    - premium
    - premiumCurrency
    - priceCurrency
    - pricePerStock
    - sellerId
    - settlementDate
    - ticker
    type: object
  handler.ExerciseContractRequest:
    properties:
      accountNumber:
        type: string
    required:
    - accountNumber
    type: object
info:
  contact: {}
  description: |-
    Bank-to-bank coordination service implementing the inter-bank
    transaction and OTC negotiation protocol. Authentication is
    per-peer via the X-Api-Key header (see §2.10).
  title: Interbank Service API
  version: "1.0"
paths:
  /api/peer-otc/contracts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PeerContract'
            type: array
      security:
      - BearerAuth: []
      summary: List my cross-bank OTC contracts
      tags:
      - peer-otc
  /api/peer-otc/contracts/{rn}/{id}/exercise:
    post:
      parameters:
      - description: Authoritative bank routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Contract id
        in: path
        name: id
        required: true
        type: string
      - description: Buyer account number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ExerciseContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PeerContract'
      security:
      - BearerAuth: []
      summary: Exercise a cross-bank OTC contract
      tags:
      - peer-otc
  /api/peer-otc/negotiations:
    get:
      description: |-
        Returns every cross-bank negotiation in which the
        authenticated user is a party.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OtcNegotiationView'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: List my cross-bank OTC negotiations
      tags:
      - peer-otc
    post:
      consumes:
      - application/json
      description: |-
        The authenticated user becomes the buyer. The request is
        forwarded to the seller's bank via §3.2; on success the
        authoritative id assigned by the peer is returned.
      parameters:
      - description: Offer terms
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePeerNegotiationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ForeignBankId'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "502":
          description: Peer unreachable or returned non-2xx
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Initiate a cross-bank OTC negotiation
      tags:
      - peer-otc
  /api/peer-otc/negotiations/{rn}/{id}:
    delete:
      description: Closes the negotiation locally and notifies the peer via §3.5.
      parameters:
      - description: Authoritative bank routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Withdraw from a cross-bank negotiation
      tags:
      - peer-otc
  /api/peer-otc/negotiations/{rn}/{id}/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authoritative bank routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      - description: Local account number
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.AcceptPeerNegotiationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PeerContract'
      security:
      - BearerAuth: []
      summary: Accept a cross-bank OTC negotiation
      tags:
      - peer-otc
  /api/peer-otc/negotiations/{rn}/{id}/counter:
    put:
      consumes:
      - application/json
      description: |-
        Posts a counter-offer whether the authenticated user is the
        buyer or the seller, notifying the opposing party's bank via §3.3.
      parameters:
      - description: Authoritative bank routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      - description: Counter terms
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CounterPeerNegotiationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Post a counter-offer on a cross-bank negotiation
      tags:
      - peer-otc
  /api/peer-otc/public-stocks:
    get:
      description: |-
        Aggregates §3.1 /public-stock results from every peer in
        the registry. Peers that fail are silently skipped.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PublicStock'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Browse public OTC stocks across all peer banks
      tags:
      - peer-otc
...
swagger: "2.0"